requirements.txt
requests
pandas
pydub
python-dotenv